pmsx003: Add specific support for PMS5003
I only tried this on a PMS5003 sensor. In the code, I special-cased the
PMS5003 model where it behaved differently from what was written.
However, it may be that the other sensor types are buggy and they in
fact all work the way my PMS5003 does. If that was the case, we could
remove some of the special handling.